adding more validation and comments.

+++ TriangleBlazor/Model/Triangle.cs
@@ -3,39 +3,45 @@
     public class Triangle
     {
         // sides
-        public int SideA;
-        public int SideB;
-        public int SideC;
+        public double SideA;
+        public double SideB;
+        public double SideC;
 
         // degrees
         public double AngleA;
         public double AngleB;
         public double AngleC;
 
-        public Triangle(int sideA = 0, int sideB = 0, int sideC = 0)
+        public Triangle(double sideA = 0, double sideB = 0, double sideC = 0)
         {
             SideA = sideA;
             SideB = sideB;
             SideC = sideC;
             CalculateDegrees();
         }
+
+        // general side validation
         public bool IsValid() => IsAB_C() && IsAC_B() && IsBC_A();
 
+        // break down to allow specific side errors for index page
         public bool IsAB_C() => ValidateSides(SideA, SideB, SideC);
         public bool IsAC_B() => ValidateSides(SideC, SideA, SideB);
         public bool IsBC_A() => ValidateSides(SideB, SideC, SideA);
 
-        private bool ValidateSides(int A, int B, int C) => A + B > C;
+        // use math to validate a side
+        private bool ValidateSides(double A, double B, double C) => A + B > C;
 
+
+        // generate a description for an angle
         public string CurrentAngleType()
         {
             // order the sides so that longest is always last
-            List<int> Sides = new() { SideA * SideA, SideB * SideB, SideC * SideC };
+            List<double> Sides = new() { SideA * SideA, SideB * SideB, SideC * SideC };
             Sides.Sort();
 
             // pre calculation
-            int ShortSum = Sides[0] + Sides[1];
-            int longestSide = Sides[2];
+            double ShortSum = Sides[0] + Sides[1];
+            double longestSide = Sides[2];
 
             // check what type of angle is it
             if (ShortSum == longestSide)
@@ -46,6 +52,7 @@
                 return "Obtuse";
         }
 
+        // once validated, create degrees
         public void CalculateDegrees()
         {
             if(IsValid())
@@ -56,12 +63,13 @@
             }
             else
             {
+                // if submit button clicks and fails validation, reset to 0
                 AngleA = 0;
                 AngleB = 0;
                 AngleC = 0;
             }
         }
-
+        
         private double ToDegree(double A, double B, double C) => Math.Round(Math.Acos( (A * A + B * B - C * C) / (2 * A * B) ) * (180d / Math.PI), 2);
 
         public string CurrentSideType()
